refactor: 컴포넌트 폴더 분리 및 export 한 곳에서 하도록 수정
