+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export { default as Button } from "./Common/Button";
+export { default as Skeleton } from "./Common/Skeleton";
+export { default as MovieList } from "./MovieList/MovieList";
+export { default as MovieListSkeleton } from "./MovieList/MovieListSkeleton";
+export { default as TopRatedMovie } from "./TopLayout/TopRatedMovie";
